Show regression summary in test plan run detailed panels
When a test plan run contains regressions, show a summary with the
test case path and the last pass / first fail kernel revisions in the
detailed panels with links to previous reports.

Note: Those links could be updated to point to the actual individual
failing test cases when the new views become available.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>